add authentication function

+++ data/storage.go
@@ -77,9 +77,12 @@ func (s *server) AddStorage(in *pb.StorageData, stream pb.Node_AddStorageServer)
 }
 
 func (s *server) DeleteStorage(_ context.Context, in *pb.StorageData) (*pb.Result, error) {
-	fmt.Println("----------DeleteVM-----")
+	log.Println("----------DeleteVM-----")
 	log.Printf("Receive VMID: %v", in.GetID())
 
+	if r, _ := db.SearchDBStorage(db.Storage{ID: int(in.GetID() % 100000)}); r.GroupID != int(in.GroupID) {
+		return &pb.Result{Status: false, Info: "NG"}, nil
+	}
 	if result := storage.DeleteStorage(in); result.Err != nil {
 		return &pb.Result{Status: false, Info: result.Info + "ErrorLog: " + fmt.Sprint(result.Err)}, nil
 	}
@@ -87,9 +90,12 @@ func (s *server) DeleteStorage(_ context.Context, in *pb.StorageData) (*pb.Resul
 }
 
 func (s *server) UpdateStorage(_ context.Context, in *pb.StorageData) (*pb.Result, error) {
-	fmt.Println("----------UpdateStorage-----")
+	log.Println("----------UpdateStorage-----")
 	log.Printf("Receive VMID: %v", in.GetID())
 
+	if r, _ := db.SearchDBStorage(db.Storage{ID: int(in.GetID() % 100000)}); r.GroupID != int(in.GroupID) {
+		return &pb.Result{Status: false, Info: "NG"}, nil
+	}
 	if result := storage.UpdateStorage(in); result.Err != nil {
 		return &pb.Result{Status: false, Info: fmt.Sprint(result.Err)}, nil
 	} else {
@@ -98,20 +104,20 @@ func (s *server) UpdateStorage(_ context.Context, in *pb.StorageData) (*pb.Resul
 }
 
 func (s *server) GetStorage(_ context.Context, in *pb.StorageData) (*pb.StorageData, error) {
-	fmt.Println("----------GetStorage-----")
+	log.Println("----------GetStorage-----")
 	log.Printf("Receive VMID: %v", in.GetID())
 
 	if result, err := db.SearchDBStorage(db.Storage{ID: int(in.GetID())}); err != nil {
 		return &pb.StorageData{}, err
 	} else {
 		return &pb.StorageData{
-			ID:      int64(result.ID),
+			ID:      uint64(result.ID),
 			Name:    result.Name,
-			GroupID: int64(result.GroupID),
-			Driver:  int32(result.Driver),
-			Mode:    int32(result.Mode),
+			GroupID: uint64(result.GroupID),
+			Driver:  uint32(result.Driver),
+			Mode:    uint32(result.Mode),
 			Path:    result.Path,
-			MaxSize: int64(result.MaxSize),
+			MaxSize: uint64(result.MaxSize),
 		}, nil
 	}
 }
@@ -126,13 +132,13 @@ func (s *server) GetAllStorage(_ *pb.Null, stream pb.Node_GetAllStorageServer) e
 		log.Println(result)
 		for _, data := range result {
 			if err := stream.Send(&pb.StorageData{
-				ID:      int64(data.ID),
+				ID:      uint64(data.ID),
 				Name:    data.Name,
-				GroupID: int64(data.GroupID),
-				Driver:  int32(data.Driver),
-				Mode:    int32(data.Mode),
+				GroupID: uint64(data.GroupID),
+				Driver:  uint32(data.Driver),
+				Mode:    uint32(data.Mode),
 				Path:    data.Path,
-				MaxSize: int64(data.MaxSize),
+				MaxSize: uint64(data.MaxSize),
 			}); err != nil {
 				return err
 			}
